move each kernel step into dedicated files
